[AIDEN] refactor: drop src/enrichment/ — move 9 files into src/pipeline/ (PR-C2)
Closes the enrichment/ vs pipeline/ duplication identified in PR-C2 scoping.
100% of pipeline→enrichment imports were single-file (signal_config.py); the
remaining files were lightly-used helpers with no semantic reason for a
separate folder.

Changes:
- Moved 9 .py files from src/enrichment/ → src/pipeline/:
  signal_config, email_verifier, query_translator, waterfall_v2,
  discovery_modes, campaign_trigger, discovery_filters, keyword_expander,
  location_expander
- Deleted src/enrichment/__init__.py and removed empty directory
- Moved tests/enrichment/* → tests/pipeline/* for parity
- Rewrote 42 import lines across 33 consumer files
- Fixed inline 'from enrichment.X' patterns in test files (sed missed)

Audit: original "7 dead files" estimate was wrong — all 9 files have live
consumers (3 files used internally by query_translator + campaign_trigger).
Pure relocation, zero deletions.

Verified:
- ruff check src/ — clean
- ruff format src/ — 458 files unchanged
- pytest signal_config consumers (11 stage/layer tests): 131 passed
- pytest moved test files: 57 passed
- pytest test_campaign_config_builder: 5 passed
- pytest test_pipeline/: 183 passed, 23 xfailed
- Net ruff delta on touched files: +0 (one I001 sed-introduced fixed)
