+++ src/orchestration/flows/batch_controller_flow.py
@@ -498,12 +498,12 @@ async def trigger_replacement_discovery_task(
             icp_config = row.icp_config or {}
 
             # Siege Waterfall v3: GMB-first discovery via Bright Data (Directive #144)
-            from src.enrichment.discovery_modes import (
+            from src.integrations.bright_data_client import get_bright_data_client
+            from src.pipeline.discovery_modes import (
                 CampaignConfig,
                 DiscoveryMode,
                 GMBFirstDiscovery,
             )
-            from src.integrations.bright_data_client import get_bright_data_client
 
             bd_client = get_bright_data_client()
             gmb_discovery = GMBFirstDiscovery(bright_data_client=bd_client)
