+++ src/pipeline/layer_2_discovery.py
@@ -23,8 +23,8 @@ from urllib.parse import urlparse
 import asyncpg
 import httpx
 
-from src.enrichment.signal_config import SignalConfigRepository
 from src.integrations.dfs_labs_client import DFSLabsClient
+from src.pipeline.signal_config import SignalConfigRepository
 from src.utils.domain_blocklist import is_blocked
 
 logger = logging.getLogger(__name__)
