+++ src/pipeline/signal_config.py
