+++ src/pipeline/stage_10_message_generator.py
@@ -18,7 +18,7 @@ from typing import Any
 
 import asyncpg
 
-from src.enrichment.signal_config import SignalConfigRepository
+from src.pipeline.signal_config import SignalConfigRepository
 from src.pipeline.stage_10_critic import CRITIC_PASS_THRESHOLD, critique_and_revise  # noqa: F401
 from src.utils.domain_blocklist import BLOCKED_DOMAINS
