+++ src/pipeline/stage_3_dfs_profile.py
@@ -21,8 +21,8 @@ from typing import Any
 
 import asyncpg
 
-from src.enrichment.signal_config import SignalConfigRepository
 from src.integrations.dfs_labs_client import DFSLabsClient
+from src.pipeline.signal_config import SignalConfigRepository
 
 logger = logging.getLogger(__name__)
