+++ tests/test_layer_2_discovery.py
@@ -8,7 +8,7 @@ from unittest.mock import ANY, AsyncMock, MagicMock, call, patch
 
 import pytest
 
-from src.enrichment.signal_config import ServiceSignal, SignalConfig
+from src.pipeline.signal_config import ServiceSignal, SignalConfig
 from src.pipeline.layer_2_discovery import (
     DiscoverySource,
     Layer2Discovery,
